+++ sdk/storage/Microsoft.Azure.WebJobs.Extensions.Storage.Blobs/tests/Listeners/SharedBlobQueueListenerFactoryTests.cs
@@ -22,7 +22,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.Storage.Blobs.Listeners
             var blobOptions = new BlobsOptions()
             {
                 MaxDegreeOfParallelism = maxDegreeOfParallelism,
-                MaxDequeueCount = maxDequeueCount,
+                PoisonBlobThreshold = maxDequeueCount,
             };
 
             // Act